+++ examples/front-end/advanced/babel-optimized/.babelrc.js
@@ -1,0 +1,6 @@
+const { babelOptimizerPlugin } = require('@graphql-codegen/client-preset');
+
+module.exports = {
+  presets: ['react-app'],
+  plugins: [[babelOptimizerPlugin, { artifactDirectory: './src/gql' }]],
+};