+++ packages/graphql-codegen-cli/src/init/index.ts
@@ -24,10 +24,15 @@ export async function init() {
   const config: Types.Config = {
     overwrite: true,
     schema: answers.schema,
-    documents: answers.targets.includes(Tags.browser) ? answers.documents : null,
+    ...(answers.targets.includes(Tags.client) ||
+    answers.targets.includes(Tags.angular) ||
+    answers.targets.includes(Tags.stencil)
+      ? { documents: answers.documents }
+      : {}),
     generates: {
       [answers.output]: {
-        plugins: answers.plugins.map(p => p.value),
+        ...(answers.targets.includes(Tags.client) ? { preset: 'client' } : {}),
+        plugins: answers.plugins ? answers.plugins.map(p => p.value) : [],
       },
     },
   };
@@ -48,7 +53,7 @@ export async function init() {
   // Emit status to the terminal
   log(`
     Config file generated at ${bold(relativePath)}
-    
+
       ${bold('$ npm install')}
 
     To install the plugins.